+++ i18n/interfaces/interfaces.go
@@ -1,13 +1,41 @@
 package i18nInterfaces
 
+import "reflect"
+
 type TranslatorI interface {
-	Key() string
-	Key2(message int, new string) string
-	New() TranslatorNewI
-	Other() string
-	Voice() string
+	Galidator() TranslatorGalidatorI
+	StatusCodes() TranslatorStatusCodesI
+	Translate(key string, optionalInputs ...[]any) string
 }
 
-type TranslatorNewI interface {
-	S(parameter int) string
+type TranslatorGalidatorI interface {
+	Example() string
+	Translate(key string, optionalInputs ...[]any) string
+}
+
+type TranslatorStatusCodesI interface {
+	InternalServerError() string
+	PageNotFound() string
+	Translate(key string, optionalInputs ...[]any) string
+}
+
+func translate(instance any, key string, optionalInputs ...[]any) string {
+	structType := reflect.TypeOf(instance)
+	inputs := []any{}
+	if len(optionalInputs) > 0 {
+		inputs = optionalInputs[0]
+	}
+
+	// Iterate over all methods of the struct
+	for i := 0; i < structType.NumMethod(); i++ {
+		method := structType.Method(i)
+		if method.Name == key {
+			reflectValues := make([]reflect.Value, len(inputs))
+			for i, v := range inputs {
+				reflectValues[i] = reflect.ValueOf(v)
+			}
+			return method.Func.Call(reflectValues)[0].String()
+		}
+	}
+	return key
 }