+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+- 🐛 fix: translator works fine now
 - 🐛 fix: bug fix on generating code
 - 🎉 feat: translator completed
 - 🎉 feat: added returnMethodInputs function